Fix `ease` export to contain only easing functions
